watchdog: rti_wdt: return 0 even when not ready for ping
The rti_wdt is a windowed watchdog which means we are not allowed to
ping the watchdog outside a certain timing window. When pinging the
watchdog too early the watchdog will trigger.

We used to return -EBUSY from the watchdogs ping callback when
pinging outside the timing window. We may ping the watchdog multiple
times in boot_entry() which then results in

Failed to enable watchdog: Device or resource busy

Avoid these messages by returning successfully even when pinged too
early.

Link: https://lore.kernel.org/r/20250317104022.1046085-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>