ARM: qemu: mask data aborts during AMBA identification
The barebox FIT image target is meant as an easy way to load barebox
second stage from other bootloaders.

Doing so with U-Boot 2025.01 compiled for qemu_arm64_defconfig
currently leads to a crash when accessing the PL061 GPIO controller
device for identification.

It seems that U-Boot (or TF-A, which wasn't included in my setup)
should have done some initial setup to allow access to the PL061.

barebox doesn't even have a driver for the peripheral and loading
barebox directly with -kernel works. In both cases, barebox runs in EL1.

Let's thus workaround the issue on the barebox side by using the new
CONFIG_ARM_AMBA_DABT_MASK option to turn the crash into an error
message.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250318082507.419785-4-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>