mci: detect RPMB partitions
Detect RPMB partitions and initialize a struct mci_part for them. We do
not actually register the RPMB partitions as we can't access them with
regular mci_do_block_op() and we would only access encrypted data anyway.

Right now we assume that only one RPMB is present in a system which
covers the 99% case. Should there be multiple RPMBs found we issue a
warning message as there is no way to specify which one shall be used,
the first one found will be used which may vary depending on probe
order.

Link: https://lore.kernel.org/r/20250319-rpmb-v2-3-a8801fb35cd4@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>