tee: optee: implement AVB named persistent values support
Android Verified Boot (AVB) 2.0 spec. revision 1.1 introduces support
for named persistent values that must be tamper evident and allows AVB
to store arbitrary key-value pairs.

While AVB itself is not implemented in barebox. this patch allows
barebox to access the persistent value store. This is useful on its
own and can also be used to test the RPMB access in barebox.

Link: https://lore.kernel.org/r/20250319-rpmb-v2-8-a8801fb35cd4@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>