Revert "ci: container: downgrade OpenRISC toolchain from 14.2.0 to 13.1.0"
This or1k issue turned out to be caused by the QEMU update[1] that came
with bumping the Debian release of the container image and not the
toolchain update.

We don't have the QEMU fix in the container yet, but we have a
workaround in commit b2a21df3b07b ("openrisc: dts: describe all 4 UARTs").

So let's revert the erroneous toolchain downgrade.

[1]: https://lore.kernel.org/all/20241203111513.402641-2-shorne@gmail.com/

Fixes: 6d4c6b01f4c4 ("ci: container: downgrade OpenRISC toolchain from 14.2.0 to 13.1.0")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.kernel.org/r/20250326123619.282897-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>