ARM: configs: replace NET_NFS command option with FS_NFS
The command implementation of NFS has seen virtually no development
since we added the filesystem implementation 12 years ago.

Unlike TFTP, NFS is a proper file system, so the command implementation
is much less useful. Just replace it directly with the file system
implementation and don't bother to reimplement the command in terms of
NFS, unless users actually need it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250402083606.2605313-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>