spi: atmel-quadspi: Updating the driver from the kernel repository
Updated driver from kernel repository to add limit checking
functionality.
This is the first step to ensure that the driver works properly
with QSPI memory chips larger than 128 MB.
In the updated version, the driver will return an error when
attempting to access memory larger than 128 MB, which will
not lead to data corruption in the rest of the chip.
As a bonus, the driver now supports the sama7g5 architecture.

Signed-off-by: Alexander Shiyan <eagle.alexander923@gmail.com>
Link: https://lore.barebox.org/20250403083236.29221-2-eagle.alexander923@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>