commands: devlookup: factor out cmd_export_val
Until we have support for capturing command output into a variable,
future commands will continue to have a -v parameter or similar for
printing the output of a command into a variable.

Add a helper for this that either prints it to the variable or to
stdout. This will allow easy migration to command output capture once we
have it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250404182345.2999795-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>