bootm: remove ability for late override of bootm.image
Overriding struct image_data::os_file at the location where we do it
now is not robust as the bootm handler is chosen by then and there
is no check if the override image is suitable for the original handler.

This can't be fixed in a straight-forward manner by moving the override
earlier as that would break appendroot for example.

To be able to do this robustly, some rework of the boot code may be in
order, so for now allow only overriding the oftree and the initrd.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250407104756.3379066-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>