bootm: move ELF handling to bootm handlers
The generic bootm code is for picking the correct OS, device tree and
initrd files. Additionally it parses the container formats like FIT and
uImage. It has no business to handle particular image formats, so push
the ELF handling back to the image handlers themselves. The resulting
diffstat is negative, so having the ELF code inside bootm doesn't really
serve a purpose. Move it out there.

Link: https://lore.barebox.org/20250407130729.2889508-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>