bootm: create generic FIT image handler
So far every architecture registers its own image handler for handling
FIT images. FIT is a container format and as such we can create a
generic handler for it. To find the correct handler for the kernel we
then have to detect the OS file type from the kernel within the fit
image instead of the original image.

For the generic FIT handler repurpose the sandbox FIT image handler. The
code therein was bogus anyway as it tried to open the FIT image which
will already be opened by the generic code. Remove that stuff and just
print an error message.

Link: https://lore.barebox.org/20250407130729.2889508-2-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>