bootm: have bootm_open_fit return -ENOSYS if support missing
Otherwise, os_type will be determined unconditionally for the FIT image
buffer. As the buffer has zero size, this leads to an os_type of
filetype_unknown. The ARM32 support has an "ARM raw image" that matches
against filetype_unknown leading eventually to a crash trying to execute
the FIT image's FDT header.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250409083023.2116661-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>