ARM: stm32mp: switch internal SoC code to hexadecimal
stm32mp_code() is used internally in SoC support code to record the
information what SoC we are running on per device tree.

By changing the information to hexadecimal, we can more easily apply
bitmasks to check if we are on STM32MP15 or STM32MP13.

Do that and while at it, rename the function as not to confuse any
out-of-tree users.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250409131920.1949534-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>