bootsource: add optional read of /chosen/bootsource
barebox currently /chosen properties like bootsource and earlycon,
although they can be useful in a second-stage boot scenario.

For example, on the STM32MP15, the TAMP register used to communicate
boot source from the BootROM is also used to override the reboot mode
prior to a warm reboot. As reboot modes are automatically cleared after
readout, this means that a chainloaded barebox will not be able to
determine the bootsource if the reboot mode driver is enabled.

To address this limitation, teach barebox to consume /chosen/bootsource.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250409131926.1949805-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>