commands: introduce simple base64 command
We have base64 functionality in barebox as that's used as formatting for
sealed blobs and JWTs. Add a simple command-line interface to the
functionality for aid during development and to simplify copying short
binary strings via the console.

We do not have code to encode in barebox, so the command only supports
decoding for now.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250410095718.2258908-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>