blspec: collect bootloader spec entries from ESP
The spec points out that spec entries should be collected from ESP, even
when XBOOTLDR partitions exist, but at a lower priority.

Implement this logic in barebox. As barebox as EFI payload systems have
an ESP, but may have the bootloader spec files in a separate partition,
we can't skip the walk through all partitions when we have an ESP like
we do for XBOOTLDR partitions.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20250414063152.2736649-7-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>