ARM: i.MX8MP: skov: decrease VDD_SOC to 850mV
We don't do dynamic voltage scaling for VDD_SOC, so the value set in
barebox may never change if the Linux PMIC driver doesn't intentionally
change it. The static configuration we had in barebox was for overdrive
mode, but the kernel DT since v6.15-rc1 now configures all clocks for
nominal mode instead as that's what the heat sink was specced for.

Nominal mode is 850mV, so configure BUCK1 RUN voltage to exactly that
value.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250414065009.2770749-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>