ARM: i.MX8MP: skov: halt startup until power is good
The 24V regulator supplying the system can indicate via a GPIO imminent
voltage loss. The board's capacitors hold enough charge to power the system
a while longer in such a state, but eventually, unless external power is
restored, the brownout detection of the PMIC will kick in.

For the span of time between voltage drop detection and PMIC brownout,
let's detect this situation and delay startup. This way, Linux can
detect the ongoing voltage loss, power down the eMMC, reboot into
barebox and barebox will delay boot as long as the problem persists.

Co-developed-by: Kim Christensen <kch@skov.dk>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250414065009.2770749-4-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>