ARM: skov-imx8mp: print initial > on every boot
putc_ll expands to nothing if CONFIG_DEBUG_LL is not defined. As we have
normal CONFIG_PBL_CONSOLE support, we can just use normal printing
functions to always output. Let's do that, so we have a very early
character (>) printed to console to indicate that barebox was entered,
even if further boot was to hang shortly after for whatever reason.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250414065009.2770749-5-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>