mux: port Linux multiplexer framework
The mux framework defines a contract where providers can provide
multiple mux states that consumers can switch between without having to
know how exactly the state switch is realized in hardware.

In addition to the C API, user interaction is possible over the shell
via the .state parameter that holds the currently
active state (or -1 if mux was not yet used). An additional .idle
parameter informs the user whether the mux has been claimed by a
consumer.

The Kconfig symbol is hidden by default and is supposed to be selected
by drivers making use of the binding or users can enable COMPILE_TEST
and make use of the shell API instead.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250414070524.2784496-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>