ARM: stm32mp: drop all images with legacy STM32 header
The legacy .stm32 images have been deprecated in TF-A v2.7 and removed
altogether in v2.8, released in 2022.

It's time that we move on in barebox as well and drop the legacy images.

Users wishing to continue to use new barebox versions need to upgrade
their TF-A versions and switch to FIP images.

For purposes of network boot, two alternatives are possible: The
barebox.fit image will contain all built device trees and externally
assembled FIT images can be booted if the new FIP bootm handler is enabled.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250414120151.3479508-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>