net: dhcp: mention that eth0 is the default
dhcp has always defaulted to eth0 when called without arguments, which
is not a good default for systems where eth0 is the management port of a
DSA switch.

Point out that eth0 is used by default when dhcp is called without
arguments and while at it make ignored positional arguments an error.

In general, users probably should use ifup -a (or -a1) instead.

Cc: Lars Schmidt <lsc@pengutronix.de>
Cc: Michael Grzeschik <m.grzeschik@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250416081646.2201914-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>