ARM: dts: tqma93xx add dummy supply to regulator
The regulator for the usdhc2 references the BUCK4 as its input.
The BUCK4 is part of the pca9451, which in turn is not a regulator
in barebox, so there is a problem during initialisation:

WARNING: imx-esdhc 42860000.mmc@42860000.of: Failed to get 'vmmc' regulator (ignored)

Since the dependency could not be resolved, the reg_usdhc2_vmmc
regulator is not correctly initialized, leaving the gpio in a floating
state, which results in the GPIO being low, which in turn disables
the SD-Card.

Signed-off-by: Fabian Pflug <f.pflug@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250417084723.3368038-1-f.pflug@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>