mci: add common PBL helper for chainloading after BootROM initialization
We have a number of PBL drivers that communicate directly with a
SD-Card or eMMC already initialized by the BootROM and put into
transfer state. This lets us skip having a stripped down MMC framework
in the prebootloader.

Instead of duplicating this code across drivers, let's put the duplicate
code into a common location.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250422052635.3423961-5-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>