mci: pbl: add autodetection of BootROM-initialized standard capacity cards
We more or less manage without driver model in PBL, because we reuse
hardware setup done by the BootROM.

In case of SD/eMMC cards, this presents one annoying problem: A card in
transfer state can't be asked whether it's standard or high capacity,
but we require this information to know whether we need to use a byte or
a block offset for the READ_MULTIPLE_BLOCK command.

We had a number of workaround for that:

  - On i.MX, the bootloader is located raw near the start of the SD/MMC,
    so we just read from offset 0 onwards and then jump to the correct
    location

  - On AT91SAM9, the BootROM supports only standard capacity SD-Cards,
    so we hardcode that

  - On older platforms, we have a separate configuration for the
    first stage that can use the normal mci-core and reinitialized
    the card with the downside of needing two builds.

  - On everything else, we assume that we have a high capacity card.

Fortunately, at the modest cost of 4 commands, we can get out
of transmission mode, ask the card for its capacity and back,
so let's do that.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250422052635.3423961-6-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>