mci: omap_hsmmc: add xload implementation for PBL
Luckily, the BootROM does no reset of hardware after handing over
control to the MLO and thus SD card, host and clock/resets are all
configured for barebox MLO to directly send MMC_CMD_READ_MULTIPLE_BLOCK.

Add a PBL driver that does this.

This will be used to enable a single barebox build to produce both MLO
and second stage images for the Beaglebone Black.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250422052635.3423961-9-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>