font: fbconsole: fix underflow when printing non-ASCII characters
With signed characters, non-ASCII characters (i.e. valued >= 128)
are negative and remain so when casted to int.

find_font_index will thus multiply the index with a negative number and
look for font data before the actual start of the font.

A cast to unsigned fixes this, because all fonts that trigger this if
clause's branch have 256 characters.

Fixes: 27f79c05ab42 ("video: implement framebuffer console")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20250422053641.3435585-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>